Extract Header from Layout
