+++ src/components/Layout/Layout.js
@@ -1,58 +1,19 @@
-import React, { useState, useContext } from 'react';
-import { AiOutlineMenu } from "react-icons/ai";
-import { useLocation, useNavigate } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-
+import React, { useState } from 'react';
 import Sidebar from '../Layout/Navigation/Sidebar';
-import logo from '../../assets/images/wallet_white.png';
 import WelcomeTourGuide from '../WelcomeTourGuide/WelcomeTourGuide';
 import BottomNav from './Navigation/BottomNav';
-import StatusContext from '../../context/StatusContext';
-import ConnectionStatusIcon from './Navigation/ConnectionStatusIcon';
+import Header from './Header';
 
 const Layout = ({ children }) => {
-	const { isOnline } = useContext(StatusContext);
-	const location = useLocation();
-	const navigate = useNavigate();
 	const [isOpen, setIsOpen] = useState(false);
 	const toggleSidebar = () => setIsOpen(!isOpen);
-	const { t } = useTranslation();
-
-	const handleNavigate = (path) => {
-		if (location.pathname === path) {
-			window.location.reload();
-		} else {
-			navigate(path);
-		}
-	};
 
 	return (
 		<div className="flex min-h-screen">
 			<Sidebar isOpen={isOpen} toggle={toggleSidebar} />
 
 			{/* Header */}
-			<header
-				className={`${isOpen ? 'hidden' : 'z-50 fixed top-0 left-0 w-full bg-primary dark:bg-primary-hover text-white flex items-center justify-between p-4 shadow-md md:hidden rounded-b-lg'}`}
-			>
-				<ConnectionStatusIcon size={25} />
-				<div className="flex items-center">
-					<button className='mr-2' onClick={() => handleNavigate('/')}>
-						<img
-							src={logo}
-							alt="Logo"
-							className="w-10 h-auto cursor-pointer"
-						/>
-					</button>
-					<a href={('/')}
-						className="text-white text-xl font-bold cursor-pointer"
-					>
-						{t('common.walletName')}
-					</a>
-				</div>
-				<button className="text-white xm:hidden" onClick={toggleSidebar}>
-					<AiOutlineMenu size={24} />
-				</button>
-			</header>
+			{!isOpen && <Header toggleSidebar={toggleSidebar} />}
 
 			<div className={`w-3/5 ${isOpen ? "hidden md:flex" : "flex"} flex-col flex-grow `}>
 				{/* Content */}
